+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Removed deprecated positional arguments for time zone and range
 - Made `last` option extend to end of current period
 - Raise error when `day_start` and `week_start` passed to unsupported methods
+- Dropped support for `mysql` gem (last release was 2013)
 
 ## 4.3.0 (2019-12-26)